+++ internal/service/repository/create_from_url.go
@@ -1,0 +1,72 @@
+package repository
+
+import (
+	"fmt"
+	"os"
+	"os/exec"
+
+	"gitlab.com/gitlab-org/gitaly/internal/command"
+	"gitlab.com/gitlab-org/gitaly/internal/helper"
+
+	pb "gitlab.com/gitlab-org/gitaly-proto/go"
+
+	"golang.org/x/net/context"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
+)
+
+func (s *server) CreateRepositoryFromURL(ctx context.Context, req *pb.CreateRepositoryFromURLRequest) (*pb.CreateRepositoryFromURLResponse, error) {
+	if err := validateCreateRepositoryFromURLRequest(req); err != nil {
+		return nil, grpc.Errorf(codes.InvalidArgument, "CreateRepositoryFromURL: %v", err)
+	}
+
+	repository := req.Repository
+
+	repositoryFullPath, err := helper.GetPath(repository)
+	if err != nil {
+		return nil, err
+	}
+
+	if _, err := os.Stat(repositoryFullPath); !os.IsNotExist(err) {
+		return nil, grpc.Errorf(codes.InvalidArgument, "CreateRepositoryFromURL: dest dir exists")
+	}
+
+	args := []string{
+		"clone",
+		"--bare",
+		"--",
+		req.Url,
+		repositoryFullPath,
+	}
+	cmd, err := command.New(ctx, exec.Command(command.GitPath(), args...), nil, nil, nil)
+	if err != nil {
+		return nil, grpc.Errorf(codes.Internal, "CreateRepositoryFromURL: clone cmd start: %v", err)
+	}
+	if err := cmd.Wait(); err != nil {
+		os.RemoveAll(repositoryFullPath)
+		return nil, grpc.Errorf(codes.Internal, "CreateRepositoryFromURL: clone cmd wait: %v", err)
+	}
+
+	// CreateRepository is harmless on existing repositories with the side effect that it creates the hook symlink.
+	if _, err := s.CreateRepository(ctx, &pb.CreateRepositoryRequest{Repository: repository}); err != nil {
+		return nil, grpc.Errorf(codes.Internal, "CreateRepositoryFromURL: create hooks failed: %v", err)
+	}
+
+	if err := removeOriginInRepo(ctx, repository); err != nil {
+		return nil, grpc.Errorf(codes.Internal, "CreateRepositoryFromURL: %v", err)
+	}
+
+	return &pb.CreateRepositoryFromURLResponse{}, nil
+}
+
+func validateCreateRepositoryFromURLRequest(req *pb.CreateRepositoryFromURLRequest) error {
+	if req.GetRepository() == nil {
+		return fmt.Errorf("empty Repository")
+	}
+
+	if req.GetUrl() == "" {
+		return fmt.Errorf("empty Url")
+	}
+
+	return nil
+}