+++ internal/service/repository/create_from_url_test.go
@@ -1,0 +1,110 @@
+package repository
+
+import (
+	"io/ioutil"
+	"os"
+	"path"
+	"testing"
+
+	"gitlab.com/gitlab-org/gitaly/internal/helper"
+	"gitlab.com/gitlab-org/gitaly/internal/testhelper"
+
+	pb "gitlab.com/gitlab-org/gitaly-proto/go"
+
+	"github.com/stretchr/testify/require"
+	"google.golang.org/grpc/codes"
+)
+
+const defaultStorageName = "default"
+
+func TestSuccessfulCreateRepositoryFromURLRequest(t *testing.T) {
+	server, serverSocketPath := runRepoServer(t)
+	defer server.Stop()
+
+	client, conn := newRepositoryClient(t, serverSocketPath)
+	defer conn.Close()
+
+	ctx, cancel := testhelper.Context()
+	defer cancel()
+
+	importedRepo := &pb.Repository{
+		RelativePath: "imports/test-repo-imported.git",
+		StorageName:  defaultStorageName,
+	}
+
+	req := &pb.CreateRepositoryFromURLRequest{
+		Repository: importedRepo,
+		Url:        "https://gitlab.com/gitlab-org/gitlab-test.git",
+	}
+
+	_, err := client.CreateRepositoryFromURL(ctx, req)
+	require.NoError(t, err)
+
+	importedRepoPath, err := helper.GetRepoPath(importedRepo)
+	require.NoError(t, err)
+	defer os.RemoveAll(importedRepoPath)
+
+	testhelper.MustRunCommand(t, nil, "git", "-C", importedRepoPath, "fsck")
+
+	remotes := testhelper.MustRunCommand(t, nil, "git", "-C", importedRepoPath, "remote")
+	require.NotContains(t, string(remotes), "origin")
+
+	info, err := os.Lstat(path.Join(importedRepoPath, "hooks"))
+	require.NoError(t, err)
+	require.NotEqual(t, 0, info.Mode()&os.ModeSymlink)
+}
+
+func TestFailedCreateRepositoryFromURLRequestDueToExistingTarget(t *testing.T) {
+	server, serverSocketPath := runRepoServer(t)
+	defer server.Stop()
+
+	client, conn := newRepositoryClient(t, serverSocketPath)
+	defer conn.Close()
+
+	ctx, cancel := testhelper.Context()
+	defer cancel()
+
+	testCases := []struct {
+		desc     string
+		repoPath string
+		isDir    bool
+	}{
+		{
+			desc:     "target is a directory",
+			repoPath: "imports/test-repo-import-dir.git",
+			isDir:    true,
+		},
+		{
+			desc:     "target is a file",
+			repoPath: "imports/test-repo-import-file.git",
+			isDir:    false,
+		},
+	}
+
+	for _, testCase := range testCases {
+		t.Run(testCase.desc, func(t *testing.T) {
+			importedRepo := &pb.Repository{
+				RelativePath: "imports/test-repo-imported.git",
+				StorageName:  defaultStorageName,
+			}
+
+			importedRepoPath, err := helper.GetPath(importedRepo)
+			require.NoError(t, err)
+
+			if testCase.isDir {
+				require.NoError(t, os.MkdirAll(importedRepoPath, 0770))
+			} else {
+				require.NoError(t, ioutil.WriteFile(importedRepoPath, nil, 0644))
+			}
+			defer os.RemoveAll(importedRepoPath)
+
+			req := &pb.CreateRepositoryFromURLRequest{
+				Repository: importedRepo,
+				Url:        "https://gitlab.com/gitlab-org/gitlab-test.git",
+			}
+
+			_, err = client.CreateRepositoryFromURL(ctx, req)
+			testhelper.AssertGrpcError(t, err, codes.InvalidArgument, "")
+		})
+	}
+}